// src/monthly-table/dto/update-monthlyTable.dto.ts
export class UpdateMonthlyTableDto {
  magasin: string;
  unites: {
    nom: string;
    matin: Array<number>;
    midi: Array<number>;
    soir: Array<number>;
    totalMatin: number;
    totalMidi: number;
    totalSoir: number;
  }[];

  totalMatin: Array<number>;
  totalMidi: Array<number>;
  totalSoir: Array<number>;
  totalRow: Array<number>;
}